rename network.py to nodes.py and move region mappers to helpers
